+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `RECEIPT_ID` option to external id generation for forms
+
 ## [2.0.2] - 2023-03-13
 
 ### Added